pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain_tutorials"
description = "Generic Minimal Data Science Application Template"
authors = [{name = "Hariharan Srinath", email = "srinathh@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langchain-text-splitters",
    "langgraph",
    "pypdf",
    "pandas",
    "python-dotenv",
    "bs4",
    "pydantic"
]
dynamic = ["version"]
license = {file = "LICENSE"}

[project.optional-dependencies]
dev = [
    "jupyter",
    "pytest",
    "pytest-cov",
    "flake8",
    "black",
]

[project.scripts]
langchain_tutorials= "app.app:main"

[tool.setuptools.packages.find]
where = ["src"]
